rename the controller file
