+++ src/tests/controllers/TextStyleController.test.ts
@@ -1,5 +1,5 @@
 import MockEditorAPI from '../__mocks__/FrameProperties';
-import { TextStyleController } from '../../controllers/TextSelectionController';
+import { TextStyleController } from '../../controllers/TextStyleController';
 
 let mockedTextProperties: TextStyleController;
